Fix Vector usage in PDBFileIO, Molecule, and Atom
Change Atom to construct its coordinates using Strings. This eliminates
the imprecision from converting the Strings in PDB files to doubles.
Vector from jMath now prefers to accept String as constructor arguments
because Strings can be built into more accurate BigDecimals.

Fix calcCenterOfMass in Molecule since a Vector now returns BigDecimals.

+++ src/main/java/com/aaronpmaus/jProt/protein/Molecule.java
@@ -50,9 +50,9 @@ public class Molecule {
         for(Node<Atom> n: atomsList){
             Atom a = n.get();
             this.mass += a.getMass();
-            comX += a.getMass()*a.getCoordinates().getValue(0);
-            comY += a.getMass()*a.getCoordinates().getValue(1);
-            comZ += a.getMass()*a.getCoordinates().getValue(2);
+            comX += a.getMass()*a.getCoordinates().getValue(0).doubleValue();
+            comY += a.getMass()*a.getCoordinates().getValue(1).doubleValue();
+            comZ += a.getMass()*a.getCoordinates().getValue(2).doubleValue();
         }
         this.centerOfMass = new Vector(comX/mass, comY/mass, comZ/mass);
     }
